% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_car_ads.R
\name{fileNameFromUrl}
\alias{fileNameFromUrl}
\title{Hash URL (for naming saved files)}
\usage{
fileNameFromUrl(url)
}
\arguments{
\item{url}{String}
}
\value{
character
}
\description{
Hash URL (for naming saved files)
}
