% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_ad_details.R
\name{getAdDetailsFromUrl}
\alias{getAdDetailsFromUrl}
\title{Get car ad details from ad URL}
\usage{
getAdDetailsFromUrl(ad_url)
}
\arguments{
\item{ad_url}{String, URL to the specific car ad page}
}
\value{
data_frame, with one raw, and columns containing the ad details
}
\description{
Get car ad details from ad URL
}
\examples{
\dontrun{
  ad_url <- paste0(
      "https://www.hasznaltauto.hu/szemelyauto/volvo/xc60/",
      "volvo_xc60_2_0_t8_twin_engine_inscription_geartronic-13821432"
  )

  getAdDetailsFromUrl(ad_url)
}

}
