% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_car_ads.R
\name{getAllCarAds}
\alias{getAllCarAds}
\title{Get all car ads from search landing page URL.}
\usage{
getAllCarAds(search_landing_url, wait_between_requests = 0)
}
\arguments{
\item{search_landing_url}{String, URL for search results.}

\item{wait_between_requests}{Integer, how many seconds to wait between sending request
to get specific details from car advertisements.}
}
\value{
data_frame, containing details for all ads found from the starting URL.
}
\description{
Get all car ads from search landing page URL.
}
\examples{
\dontrun{
  getAllCarAds("https://www.hasznaltauto.hu/szemelyauto/volvo/xc60")
}

}
