% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_car_ads.R
\name{getAllCarAdUrls}
\alias{getAllCarAdUrls}
\title{Get all car ad URLs from search landing page URL}
\usage{
getAllCarAdUrls(search_landing_url)
}
\arguments{
\item{search_landing_url}{String, URL for search results.}
}
\value{
character vector, containing the url to all car ads relevant to search
}
\description{
Get all car ad URLs from search landing page URL
}
