% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasznaltautoR.R
\docType{package}
\name{hasznaltautoR}
\alias{hasznaltautoR}
\alias{hasznaltautoR-package}
\title{\code{hasznaltautoR} package}
\description{
Easy scraping of car ads from \href{https://www.hasznaltauto.hu/}{hasznaltauto.hu}
}
\details{
See the README on
\href{https://github.com/tomiaJO/hasznaltautoR#readme}{GitHub}
}
