% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_car_ads.R
\name{saveAdsToCSV}
\alias{saveAdsToCSV}
\title{Save ads resulting for scraping to a .csv file}
\usage{
saveAdsToCSV(ads, search_landing_url, save_path)
}
\arguments{
\item{ads}{String}

\item{search_landing_url}{String}

\item{save_path}{String}
}
\value{
Nothing, saves file to location.
}
\description{
Save ads resulting for scraping to a .csv file
}
